Fix event_page toggle
